fixed bug where time in microseconds was stored in a uint32 where it couldn't possibly fit.
i'll need to verify if this cures the issue i've experienced, but it is very likely to be the cause. i've suspected the jumps in ARXStartTime are valid jumps and simply the intended function (although insane) of the original code. the assertions have not been triggered yet in testing, but when breaking due to an invalid time value i did notice this uint32 issue. a copy-paste bug?

+++ src/platform/Time.cpp
@@ -72,7 +72,7 @@ u64 getUs() {
 	LARGE_INTEGER counter;
 	QueryPerformanceCounter(&counter);
 	// Ugly trick to avoid losing precision...
-	u32 valUs = (counter.QuadPart * 1000) / (FREQUENCY_HZ / 1000);
+	u64 valUs = (counter.QuadPart * 1000) / (FREQUENCY_HZ / 1000);
 	return valUs;
 }
